+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 - Added headers parameter to ipwhois.Net.get_http_json() (issue #98).
 - Fixed ASN HTTP lookup (fallback) Accept headers (issue #98).
 - Fixed HTTP decoding, set to utf-8 (italomaia)
+- IPWhois.lookup() deprecated (#96), and will be removed in a future release
+  (TBD). Use IPWhois.lookup_whois() instead.
 
 0.11.2 (2016-02-25)
 -------------------